Pulls Sublime, Slack, and Rocket.Chat roles into 1
NOTE: This commit adds some funky behavior where a portion of the roles get re-
applied despite already having been applied. I think this is a result of a bug
in the ansible code where including a file containing tasks and looping over it
using with_items results in errant code execution. It's strange because I'd
have expected the code to be executed three times, since I'm installing
Sublime, Slack, and Rocket.Chat+, but the code only seems to execute twice for
a portion.

Oh wells,

¯\_(ツ)_/¯